Moved functions to new file
